CustomerRelationsWebsite/requirements.txt
asgiref==3.5.1
autopep8==1.6.0
Django
IMAPClient==2.2.0
pycodestyle==2.8.0
six==1.16.0
sqlparse
toml==0.10.2
tzdata==2022.1

python-dotenv

whitenoise
gunicorn
django-debug-toolbar

celery
redis
hiredis

django
django-celery-results
djangorestframework
django-api-generator
drf-spectacular
django-environ
django-allauth
django-sslserver
dj-database-url
django-mongodb-engine
django-simple-history
django-import-export
djongo
# django-admin-volt
django-filter
drf-spectacular

timestamps
tablib
Pillow
asgiref
autopep8
pycodestyle
pytz
sqlparse
toml
Unipath
pandas
numpy
